79 refactor + optimize

+++ src/index.ts
@@ -328,13 +328,19 @@ app.whenReady()
         })
 
         ipcMain.handle('show-image', async (e, filePath) => {
-            const extension = path.extname(filePath)
-            const img = fs.readFileSync(filePath).toString('base64')
-            return `data:image/${extension};base64,${img}`
+            if (fs.existsSync(filePath)) {
+                const extension = path.extname(filePath)
+                const img = fs.readFileSync(filePath).toString('base64')
+                return `data:image/${extension};base64,${img}`
+            }
+
+            return ''
         })
 
         ipcMain.handle('delete-file', async (e, path) => {
-            fs.rmSync(path)
+            if (fs.existsSync(path)) {
+                fs.rmSync(path)
+            }
         })
     })
     .catch(() => null)
